distro/rhel7: move imageType implementation to separate file
